refactor: split drawer scripts into other file
